requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.2
aiologger==0.7.0
aiosignal==1.3.2
alphacast==0.1.8.5
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
Brotli==1.1.0
browserforge==1.2.3
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
discord-webhook==1.4.1
distlib==0.3.9
filelock==3.18.0
frozenlist==1.6.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
multidict==6.4.4
numpy==2.2.6
orjson==3.10.18
pandas==2.2.3
platformdirs==4.3.8
propcache==0.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
yarl==1.20.0
